implement virtual nodes

+++ misc/sql/storage/migration.sql
@@ -3,7 +3,8 @@ create table files
     name varchar not null
         primary key
         unique,
-    size int,
+    size bigint,
     data bytea   not null
 );
-
+-- alter system set log_statement to 'all';
+-- select pg_reload_conf();
